+++ shared/src/commonMain/kotlin/site/remlit/snowdrop/view/settings/AboutSnowdropView.kt
@@ -1,5 +1,7 @@
 package site.remlit.snowdrop.view.settings
 
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.FlowRow
@@ -15,6 +17,11 @@ import androidx.compose.material3.OutlinedButton
 import androidx.compose.material3.Text
 import androidx.compose.material3.TopAppBar
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalUriHandler
@@ -24,9 +31,11 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import org.jetbrains.compose.resources.painterResource
 import org.jetbrains.compose.resources.stringResource
+import site.remlit.snowdrop.DebugRoute
 import site.remlit.snowdrop.GradleVariables
 import site.remlit.snowdrop.component.ViewSurface
 import site.remlit.snowdrop.util.LocalNavController
+import site.remlit.snowdrop.util.navigationBarInteractionSource
 import site.remlit.snowdrop.util.translation
 import snowdrop.shared.generated.resources.Res
 import snowdrop.shared.generated.resources.about_snowdrop
@@ -42,6 +51,14 @@ import snowdrop.shared.generated.resources.view_source
 @Composable
 fun AboutSnowdropView() = ViewSurface {
 	val navHandler = LocalNavController.current
+
+	var versionClicks by remember { mutableStateOf(0) }
+	LaunchedEffect(versionClicks) {
+		if (versionClicks == 5) {
+			navHandler.navigate(DebugRoute)
+			versionClicks = 0
+		}
+	}
 
 	TopAppBar(
 		navigationIcon = {
@@ -75,7 +92,12 @@ fun AboutSnowdropView() = ViewSurface {
 		)
 
 		Column(
-			horizontalAlignment = Alignment.CenterHorizontally
+			horizontalAlignment = Alignment.CenterHorizontally,
+			modifier = Modifier.clickable(
+				interactionSource = MutableInteractionSource(),
+				indication = null,
+				onClick = { versionClicks++ }
+			)
 		) {
 			Text(
 				"${GradleVariables.version} (${GradleVariables.gitCommit}@${GradleVariables.gitBranch})",