+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - StatusMediaAttachment component for reusing the same logic across status media previews and the full viewer
 - transitionedComposable function which automatically applies correct transitions to the page
 - Notification filtering selection in notifications page
+- Remove debug button in settings, instead open debug after clicking the version information on the About Snowdrop page 5 times
 
 ### Fixed
 - AMOLED theme has better contrast on some surfaces